Update Enemy Sprite and Add ground for game
The ground surface has been included in game.
Player will be able to stand onto the ground properly.
Update the enemies' sprite properly with transparent sprite.
Proper check for player and enemies are still in progress, even there are many solutions are included, but it has not been working yet. This task is in progress for now.
Add a new property in getTransform class, SDL_FRect.
Add this new property to check for platform.
Trying to get it work until the next update.