Update AI with proper range for LOS (#35)
When the player is in the proper field of view of enemies, then the enemies will shoot bullet.
Problem: since the enemies are in the scrolling chain, if the player get to the back of the enemies by jumping, the enemies will turn back, but it will be unable to shoot because the subsequent chain of random spawning obstacles will keep to check for the next ones (obstacle or enemies). This may result in ignorance for checking shooting properly.