beta (#51)
* Create test.txt

* More Obstacles Added

Added a fireball and spiky enemies

* More Obstacles Added (#25)

Added a fireball and spiky enemies

* audio files

* Fixed bullet position

* Expand window size to 1152*768, updated suitable ground file, use the window size as a variable to position the ui, ScoreBoard class added, deleted Next button

* Update Jumping with parapolic landing

* Update Jumping with parapolic landing (#27)

* Animated alien

* Animated Alien

* Update spawning bullet for enemy

Enemy will be able to shoot on both sides for now whenever the players are in the range of the enemy.

* Create test.txt

* fixed minor merge errors

scrolling is still broken

* scrolling now works

-- shooting has also been added but the enemy can see the player from far away, we should try to decrease to alien's LOS

* Update AI with proper range for LOS

When the player is in the proper field of view of enemies, then the enemies will shoot bullet.
Problem: since the enemies are in the scrolling chain, if the player get to the back of the enemies by jumping, the enemies will turn back, but it will be unable to shoot because the subsequent chain of random spawning obstacles will keep to check for the next ones (obstacle or enemies). This may result in ignorance for checking shooting properly.

* Enemy sprites,and the grounnd is updated

* Enemy sprites,and the grounnd is updated (#36)

Co-authored-by: miarai25 <60987674+miarai25@users.noreply.github.com>

* Update AI with proper range for LOS (#35)

When the player is in the proper field of view of enemies, then the enemies will shoot bullet.
Problem: since the enemies are in the scrolling chain, if the player get to the back of the enemies by jumping, the enemies will turn back, but it will be unable to shoot because the subsequent chain of random spawning obstacles will keep to check for the next ones (obstacle or enemies). This may result in ignorance for checking shooting properly.

* Enemy sprites,and the grounnd is updated (#37)

Co-authored-by: Miaraii25 <65626916+Miaraii25@users.noreply.github.com>
Co-authored-by: miarai25 <60987674+miarai25@users.noreply.github.com>

* Level Select Scene added, Pause and Continue button deleted, Exit Button added

* Level Select Scene added, Pause and Continue button deleted, Exit button added

* Update for clean function

* spikes working

* more updates to kill player

* deleted test files

* Merge pull request #5 from CLBaziw/master (#40)

* Merge pull request #5 from CLBaziw/master (#41)

* Merge pull request #5 from CLBaziw/master (#42)

* Merge pull request #5 from CLBaziw/master (#43)

* predator and boss sprites

* predator and boss sprites (#44)

* Enemy sprites,and the grounnd is updated

* predator and boss sprites

Co-authored-by: miarai25 <60987674+miarai25@users.noreply.github.com>

* ground,spike,scroll

created a new ground for level 2,scrolled the grounds and background,also reduced the size of the spikes

* spike,ground,scroll (#45)

* Enemy sprites,and the grounnd is updated

* predator and boss sprites

* ground,spike,scroll

created a new ground for level 2,scrolled the grounds and background,also reduced the size of the spikes

Co-authored-by: miarai25 <60987674+miarai25@users.noreply.github.com>

* Add different types of enemy

Update the bullet type by using switch statement
Add another type of enemy: predator
Error: after the obstacles are spawning for a while, it will crash. This error has been existing before in the previous release and not been improved.

* Lever 3 Button added, Added Score to End Screen

* resolved crashing errors on player death

* trying to resolve merge errors

* HealthBar,ScoreBoard and collision with spike

* HealthBar,spike,Score

* Add special skill for Player

Now press X for shooting.
After firing a certain number of bullet, the player gonna accummulate the energy to shoot a special bullet.

* fixed merge errors from updated bullet

* removed active sprite list from object pool

* updated enemy bullets

made enemy shoot function more efficient

* made score into a getter and setter

Co-authored-by: predator1245 <jashshah37@yahoo.com>
Co-authored-by: psmyang <65626707+psmyang@users.noreply.github.com>
Co-authored-by: Miaraii25 <65626916+Miaraii25@users.noreply.github.com>
Co-authored-by: JohnnyDiep1021 <60988318+JohnnyDiep1021@users.noreply.github.com>
Co-authored-by: BertMarsh <60789232+BertMarsh@users.noreply.github.com>
Co-authored-by: predator1245 <61018687+predator1245@users.noreply.github.com>
Co-authored-by: psmyang <ymk1234567890ymk@sina.com>
Co-authored-by: Johnny Diep <longdiep10102001@gmail.com>
Co-authored-by: miarai25 <60987674+miarai25@users.noreply.github.com>
Co-authored-by: miarai25 <Mariam.Ogunlesi@georgebrown.ca>